+++ src/dfxvm/install.rs
@@ -33,11 +33,15 @@ use std::path::{Path, PathBuf};
 use tar::Archive;
 use tempdir::TempDir;
 
+pub fn installed(version: &Version, locations: &Locations) -> bool {
+    locations.version_dir(version).exists()
+}
+
 pub async fn install(version: Version) -> Result<(), InstallError> {
     let locations = Locations::new()?;
     let settings = Settings::load_or_default(&locations.settings_path())?;
     let version_dir = locations.version_dir(&version);
-    if version_dir.exists() {
+    if installed(&version, &locations) {
         info!("dfx {version} is already installed");
         return Ok(());
     }