+++ tests/suite/common/settings.rs
@@ -1,3 +1,4 @@
+use itertools::Itertools;
 use serde_json::{json, Value};
 use std::fs::create_dir_all;
 use std::path::PathBuf;
@@ -9,6 +10,20 @@ pub struct Settings {
 impl Settings {
     pub fn new(path: PathBuf) -> Self {
         Self { path }
+    }
+
+    pub fn read_default_version(&self) -> String {
+        self.read()["default_version"].as_str().unwrap().to_string()
+    }
+
+    pub(crate) fn sorted_keys(&self) -> Vec<String> {
+        self.read()
+            .as_object()
+            .unwrap()
+            .keys()
+            .sorted()
+            .map(|s| s.to_string())
+            .collect()
     }
 
     pub fn write_default_version(&self, version: &str) {